Added admin queries
Added isBlocked property for user model.
Added routes for toggle isAdmin, isBlocked user properties.
Added route for user delete.

Added check for access in routes with necessary authorizations.

Improved services.

Added checkAdmin, checkBlock middlewares.

+++ controllers/chapter.js
@@ -1,24 +1,30 @@
-const Fanfic = require("../models/fanfic");
 const Chapter = require("../models/chapter");
 
+const fanficService = require("../services/fanficService");
 const likeService = require("../services/likeService");
 
 exports.create = async (req, res, next) => {
   try {
+    const { _id: authId } = req.userData;
     const { name, content, fanfic } = req.body;
-    const chapterCount = await Chapter.estimatedDocumentCount({ fanfic });
-    const chapter = await Chapter.create({
-      fanfic,
-      name,
-      content,
-      number: chapterCount + 1
-    });
-    await Fanfic.findByIdAndUpdate(chapter.fanfic, { lastUpdate: Date.now() });
-    return res.status(200).json({
-      message: "Chapter added.",
-      chapter,
-      isLiked: null
-    });
+    const isOwner = await fanficService.isOwner(authId, fanfic);
+    if (isOwner) {
+      const chapterCount = await Chapter.estimatedDocumentCount({ fanfic });
+      const chapter = await Chapter.create({
+        fanfic,
+        name,
+        content,
+        number: chapterCount + 1
+      });
+      await fanficService.fanficLastUpdateNow(fanfic);
+      return res.status(200).json({
+        message: "Chapter added.",
+        chapter,
+        isLiked: null
+      });
+    } else {
+      throw new Error("No access rights.");
+    }
   } catch (err) {
     next(err);
   }
@@ -26,7 +32,7 @@ exports.create = async (req, res, next) => {
 
 exports.getChapter = async (req, res, next) => {
   try {
-    const user = req.userData?._id;
+    const user = req.userData._id;
     const chapterId = req.params.chapterId;
     const chapter = await Chapter.findById(chapterId);
     if (!chapter) {
@@ -64,24 +70,31 @@ exports.getChapterUpdate = async (req, res, next) => {
 
 exports.update = async (req, res, next) => {
   try {
+    const { _id: authId, isAdmin } = req.userData;
     const chapterId = req.params.chapterId;
     const { name, content } = req.body;
     const chapter = await Chapter.findById(chapterId);
     if (!chapter) {
       throw new Error("Chapter not found.");
     }
-    if (name) {
-      chapter.name = name;
+    const isOwner = await fanficService.isOwner(authId, chapter.fanfic);
+    if (isAdmin || isOwner) {
+      if (name) {
+        chapter.name = name;
+      }
+      if (content) {
+        chapter.content = content;
+      }
+      chapter.lastUpdate = Date.now();
+      await chapter.save();
+      await fanficService.fanficLastUpdateNow(chapter.fanfic);
+      return res.status(200).json({
+        message: "Successful chapter update."
+      });
+    } else {
+      throw new Error("No access rights.");
     }
-    if (content) {
-      chapter.content = content;
-    }
-    chapter.lastUpdate = Date.now();
-    await chapter.save();
-    await Fanfic.findByIdAndUpdate(chapter.fanfic, { lastUpdate: Date.now() });
-    return res.status(200).json({
-      message: "Successful chapter update."
-    });
+
   } catch (err) {
     next(err);
   }
@@ -89,26 +102,33 @@ exports.update = async (req, res, next) => {
 
 exports.move = async (req, res, next) => {
   try {
+    const { _id: authId, isAdmin } = req.userData;
     const chapterId = req.params.chapterId;
     const { number } = req.body;
     const movingChapter = await Chapter.findById(chapterId);
     if (!movingChapter) {
       throw new Error("Chapter not found.");
     }
-    const swapingChapter = await Chapter.findOne({ fanfic: movingChapter.fanfic, number });
-    if (!swapingChapter) {
-      throw new Error("Chapter not found.");
+    const isOwner = await fanficService.isOwner(authId, movingChapter.fanfic);
+    if (isAdmin || isOwner) {
+      const swapingChapter = await Chapter.findOne({ fanfic: movingChapter.fanfic, number });
+      if (!swapingChapter) {
+        throw new Error("Chapter not found.");
+      }
+      swapingChapter.number = movingChapter.number;
+      swapingChapter.lastUpdate = Date.now();
+      await swapingChapter.save();
+      movingChapter.number = number;
+      movingChapter.lastUpdate = Date.now();
+      await movingChapter.save();
+      await fanficService.fanficLastUpdateNow(movingChapter.fanfic);
+      return res.status(200).json({
+        message: "Successful chapter move."
+      });
+    } else {
+      throw new Error("No access rights.");
     }
-    swapingChapter.number = movingChapter.number;
-    swapingChapter.lastUpdate = Date.now();
-    await swapingChapter.save();
-    movingChapter.number = number;
-    movingChapter.lastUpdate = Date.now();
-    await movingChapter.save();
-    await Fanfic.findByIdAndUpdate(movingChapter.fanfic, { lastUpdate: Date.now() });
-    return res.status(200).json({
-      message: "Successful chapter move."
-    });
+
   } catch (err) {
     next(err);
   }
@@ -162,15 +182,23 @@ exports.unlike = async (req, res, next) => {
 
 exports.delete = async (req, res, next) => {
   try {
+    const { _id: authId, isAdmin } = req.userData;
     const chapterId = req.params.chapterId;
-    const chapter = await Chapter.findByIdAndDelete(chapterId);
+    const chapter = await Chapter.findById(chapterId);
     if (!chapter) {
       throw new Error("Fanfic not found.");
     }
-    await Fanfic.findByIdAndUpdate(chapter.fanfic, { lastUpdate: Date.now() });
-    return res.status(200).json({
-      message: "Successful chapter delete."
-    });
+    const isOwner = await fanficService.isOwner(authId, chapter.fanfic);
+    if (isAdmin || isOwner) {
+      await Chapter.findByIdAndDelete(chapterId);
+      await likeService.removeChapterLikes(chapterId);
+      await fanficService.fanficLastUpdateNow(chapter.fanfic);
+      return res.status(200).json({
+        message: "Successful chapter delete."
+      });
+    } else {
+      throw new Error("No access rights.");
+    }
   } catch (err) {
     next(err);
   }
